requirements.txt
Django==1.7.5
Markdown==2.6
django-bootstrap3==5.1.1
django-filter==0.9.2
django-oauth2-provider==0.2.6.1
django-safedelete==0.2.1
djangorestframework==3.0.5
django-mptt==0.6.1
phonenumbers==7.0.2
mysqlclient==1.3.6
Pillow==2.7.0
